Introduces XSD schema loading and validation wrappers
In terms of level of abstraction of the API, it is kept as low as
possible without exposing intermediate structures which are required to
exist somewhere while validation process is in progress. Which is mostly
the Schema, wrapping over xmlSchemaPtr.

There are still improvements that can be made over some error handling
which will be coming as this module gets used in a downstream project.