Enable huge files processing
For security reasons [1], the libxml2 library has a hardcoded limit of 10Mb for
processing files.
The library offers the `XML_PARSE_HUGE` entry as part of the `xmlParserOption` enum
to allow programmers to override this limit.
This option can be given to the `xmlReadMemory` function to allow the processing
of files over the 10Mb limit.

This patch adds the `huge` option to the `ParserOptions` struct to allow rust-libxml
parse with options functions to use this feature.

[1] https://cve.mitre.org/cgi-bin/cvename.cgi?name=CVE-2008-4226