Protect `xmlSchemaInitTypes` from thread-unsafe behaviour.
`xmlSchemaParse` calls `xmlSchemaInitTypes`.
`xmlSchemaInitTypes` is a lazy function which is only intended to be
called once for optimization purposes - but libxml2 doesn't do this
in a thread-safe manner.  We wrap the call in a OnceLock so that it
only ever needs to be invoked once - and will do it in a thread-safe
way.